backend/requirements.txt
asgiref == 3.8.1
Django == 5.1.6
django-cors-headers == 4.7.0
djangorestframework == 3.15.2
djangorestframework-simplejwt == 5.4.0
PyJWT == 2.10.1
pytz == 2025.1
sqlparse == 0.5.3
psycopg2-binary == 2.9.10
python-dotenv == 1.0.1
celery == 5.4.0
redis == 5.2.1
channels[daphne] == 4.2.0
channels-redis == 4.2.1
django-celery-results == 2.5.1
django-celery-beat == 2.7.0
django-redis == 5.4.0
cloudinary == 1.42.2
django-watson==1.6.3
pillow == 11.1.0
pytest==8.1.1
pytest-django==4.8.0
pytest-asyncio==0.23.5
factory-boy==3.3.0
coverage==7.5.0